fix: resolve CI architecture test violations
- Add barrel export (index.ts) for registration feature module
- Move LoginPage to src/pages/login.tsx to reduce App.tsx size
- Update size.test.ts allowlist: App.tsx 572 -> 383 lines
