+++ frontend/src/features/registration/index.ts
@@ -1,0 +1,1 @@
+export { RegisterPage } from './pages/register-page'
